+++ src/components/Home.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Home = () => {
+  return (
+    <div style={{ height: "100vh", backgroundColor: "#f7f0f6" }}>
+      Decentralized Exchange System!
+    </div>
+  );
+};
+
+export default Home;
